Implementata la stampa tramite le nuove classi di WPF (prima avevo abbozzato qualcosa tramite windows.drawing che per? volevo deprecare).
Refactoring anche per quanto riguarda la dispose dei servizi. C'erano delle cose
che non mi tornavano. Ho cercato di migliorare la usabilit?.

+++ Digiphoto.Lumen.Core/src/Imaging/Immagine.cs
@@ -21,15 +21,16 @@ namespace Digiphoto.Lumen.Imaging {
 
 
 		#region Metodi astratti dall'interfaccia
-		public abstract long ww {
+		public abstract int ww {
 			get;
 		}
 
-		public abstract long hh {
+		public abstract int hh {
 			get;
 		}
 
 		public abstract void Dispose();
+
 		#endregion
 	}
 }
